+++ app/controllers/admin/slide_reconvert_controller.rb
@@ -3,7 +3,7 @@ module Admin
     before_action :set_slide, only: [:show]
 
     def show
-      @slide.send_convert_request
+      SlideConvertService.new(@slide).send_request
       redirect_to admin_slides_path, notice: t(:send_convert_request_completed)
     end
