+++ app/services/slide_file_service.rb
@@ -1,0 +1,10 @@
+class SlideFileService
+  def initialize(slide)
+    @slide = slide
+  end
+
+  def delete_files
+    CloudConfig::SERVICE.delete_slide(@slide.object_key)
+    CloudConfig::SERVICE.delete_generated_files(@slide.object_key)
+  end
+end
